// frontend/src/lib/api/generated/models/MyEventItem.ts
export class MyEventItem {
    'id': number;
    'slug': string;
    'organizationName': string;
    'name': string;
    'startAt': Date;
    'endAt': Date;
    'applicationStartAt': Date;
    'applicationEndAt': Date;
    'totalTicketNumber': number;
    'coverImageUrl': string;
    'organizeAddress'?: string | null;
    'organizeCityCode'?: string | null;
    'meetingToolCode'?: EventMeetingToolCode | null;
    'ticketName'?: string | null;
    'isOnline'?: boolean | null;
    'isOffline'?: boolean | null;
    'isApplied'?: boolean | null;
    'isBookmarked'?: boolean | null;
    'publishedAt': Date;
    'tags'?: Array<TagItem>;
    'transactionHistories'?: Array<MyTicketTransaction>;
    'minTicketPrice'?: number | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "slug",
            "baseName": "slug",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizationName",
            "baseName": "organization_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "startAt",
            "baseName": "start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "endAt",
            "baseName": "end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationStartAt",
            "baseName": "application_start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationEndAt",
            "baseName": "application_end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "totalTicketNumber",
            "baseName": "total_ticket_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "coverImageUrl",
            "baseName": "cover_image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizeAddress",
            "baseName": "organize_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "organizeCityCode",
            "baseName": "organize_city_code",
            "type": "string",
            "format": ""
        },
        {
            "name": "meetingToolCode",
            "baseName": "meeting_tool_code",
            "type": "EventMeetingToolCode",
            "format": ""
        },
        {
            "name": "ticketName",
            "baseName": "ticket_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "isOnline",
            "baseName": "is_online",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isOffline",
            "baseName": "is_offline",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isApplied",
            "baseName": "is_applied",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "isBookmarked",
            "baseName": "is_bookmarked",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "publishedAt",
            "baseName": "published_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "tags",
            "baseName": "tags",
            "type": "Array<TagItem>",
            "format": ""
        },
        {
            "name": "transactionHistories",
            "baseName": "transaction_histories",
            "type": "Array<MyTicketTransaction>",
            "format": ""
        },
        {
            "name": "minTicketPrice",
            "baseName": "min_ticket_price",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return MyEventItem.attributeTypeMap;
    }

    public constructor() {
    }
}